[UIU-2852] - Fixed remove user affiliation feature
