+++ src/main/java/org/folio/consortia/exception/ConsortiumClientException.java
@@ -2,7 +2,7 @@ package org.folio.consortia.exception;
 
 public class ConsortiumClientException extends RuntimeException {
 
-  public ConsortiumClientException(String msg, Exception exception) {
-    super(msg, exception);
+  public ConsortiumClientException(Exception exception) {
+    super(exception);
   }
 }
