+++ src/main/java/org/folio/consortia/service/impl/UserTenantServiceImpl.java
@@ -12,6 +12,7 @@ import org.folio.consortia.domain.dto.UserTenant;
 import org.folio.consortia.domain.dto.UserTenantCollection;
 import org.folio.consortia.domain.entity.TenantEntity;
 import org.folio.consortia.domain.entity.UserTenantEntity;
+import org.folio.consortia.exception.ConsortiumClientException;
 import org.folio.consortia.exception.PrimaryAffiliationException;
 import org.folio.consortia.exception.ResourceNotFoundException;
 import org.folio.consortia.repository.UserTenantRepository;
@@ -32,8 +33,6 @@ import java.util.List;
 import java.util.Objects;
 import java.util.Optional;
 import java.util.UUID;
-import java.util.regex.Matcher;
-import java.util.regex.Pattern;
 
 
 /**
@@ -254,21 +253,7 @@ public class UserTenantServiceImpl implements UserTenantService {
       log.info("User with userId {} does not exist in schema, going to use new one", userId);
       return new User();
     } catch (FeignException e) {
-      int errorCode = e.status(); // get the HTTP error code from the exception
-      String message = e.getMessage();
-      String permission = extractPermissionFromErrorMessage(message); // extract the permission from the error message
-      String formattedMessage = String.format("Access for user '%s' requires permission: %s [%d %s]", userId.toString(), permission, errorCode, message);
-      throw new IllegalStateException(formattedMessage);
-    }
-  }
-
-  private String extractPermissionFromErrorMessage(String errorMessage) {
-    Pattern pattern = Pattern.compile("permission: (\\S+)"); // regex to extract the permission from the error message
-    Matcher matcher = pattern.matcher(errorMessage);
-    if (matcher.find()) {
-      return matcher.group(1);
-    } else {
-      return "unknown";
+      throw new ConsortiumClientException(e);
     }
   }
